refactor(core): CHECKOUT-4455 Stop resolving promise with load event
BREAKING CHANGE: Previously, when a script is loaded, we return the
associated event object inside a promise object to the caller. With this
change, we return an empty promise object instead. This change is
necessary because we can't guarantee that scripts can be loaded in the
same way across different browsers. For example, some browsers don't
support `rel="preload"`, so as a fallback, we have to preload scripts
using regular XHR calls. In that case, we don't have an event object to
return to the caller. We could potentially mock the event object to keep
the return values consistent. But considering it is not a very useful
thing to return, we've decided to make a breaking change and return
nothing instead.